requirements.txt
# Zava Insurance Claims Processing - Dependencies
a2a-sdk>=0.3.0
mcp>=1.9.4
mcp[cli]>=1.0.0
fastmcp>=0.3.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-communication-email>=1.0.0
azure-ai-agents>=1.1.0b1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
aiohttp>=3.9.1
httpx>=0.27.0
semantic-kernel>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
colorama>=0.4.6
rich>=13.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
websockets>=11.0.0
orjson>=3.8.0
python-dateutil>=2.8.0
requests>=2.28.0
psutil>=5.9.0
azure-storage-blob>=12.0.0
